safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A1_T3-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Cannot use the undeclared label 'LABEL'.
S12.7_A1_T3

PASS Expected parsing failure

TEST COMPLETE
